feat: add confirmation of sending form
