+++ src/app/styles/App.scss
@@ -35,3 +35,353 @@ input {
   margin: 0 auto;
   padding: 0 20px;
 }
+
+.control {
+  --green: #32936f;
+  --black: #121314;
+  --grey: #dde2e5;
+  --primary: #5570f1;
+  --danger: #f57e77;
+  --black-5: #2b2f32;
+  --black-4: #5e6366;
+  --black-3: #83898c;
+  --black-2: #abafb1;
+  --black-1: #cfd3d4;
+  --input-disabled: #f2f4f5;
+
+  .control__container {
+    padding: 50px 0;
+  }
+
+  &__title {
+    text-align: center;
+    margin-bottom: 40px;
+  }
+
+  &__form {
+    display: flex;
+    flex-direction: column;
+    gap: 40px;
+    width: 100%;
+    max-width: 600px;
+    margin: 0 auto;
+  }
+}
+
+.input-wrapper {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  border: 1px solid var(--black-1);
+  border-radius: 8px;
+  padding: 8px 16px;
+
+  button {
+    position: absolute;
+    top: 50%;
+    right: 16px;
+    transform: translate(0, -50%);
+    border: none;
+    color: var(--black-4);
+    background: transparent;
+    cursor: pointer;
+  }
+
+  &:has(input:focus-visible),
+  &:has(input:not(:placeholder-shown)) {
+    border: 1px solid var(--black-4);
+  }
+
+  label {
+    font-weight: 400;
+    font-size: 12px;
+    color: var(--black-4);
+  }
+
+  input {
+    border: none;
+    font-weight: 400;
+    font-size: 16px;
+    color: var(--black-4);
+    -webkit-appearance: none;
+    appearance: none;
+
+    &::placeholder {
+      font-weight: 400;
+      font-size: 16px;
+      color: var(--black-2);
+    }
+
+    &:focus-visible {
+      outline: none;
+    }
+  }
+
+  input[type="number"] {
+    -moz-appearance: textfield;
+    appearance: textfield;
+  }
+
+  input::-webkit-outer-spin-button,
+  input::-webkit-inner-spin-button {
+    -webkit-appearance: none;
+    margin: 0;
+  }
+
+  &_file {
+    border: 1px solid var(--black-1);
+    align-items: flex-start;
+    font-weight: 400;
+    font-size: 12px;
+    color: var(--black-4);
+
+    input[type="file"] {
+      display: none;
+    }
+
+    &:has(input:not(:placeholder-shown)) {
+      border: 1px solid var(--black-1);
+    }
+
+    label {
+      width: auto;
+      border: 1px solid var(--black-4);
+      border-radius: 12px;
+      padding: 10px 12px;
+      font-weight: 400;
+      font-size: 14px;
+      color: var(--black-4);
+      cursor: pointer;
+    }
+  }
+
+  &.has-error,
+  &.has-error:has(input:focus-visible),
+  &.has-error:has(input:not(:placeholder-shown)) {
+    border: 1px solid var(--danger);
+
+    label {
+      color: var(--danger);
+    }
+
+    .input-error {
+      display: block;
+    }
+  }
+
+  &.is-valid {
+    border: 1px solid var(--green);
+    label {
+      color: var(--green);
+    }
+  }
+}
+.input-error {
+  display: none;
+  position: absolute;
+  top: calc(100% + 2px);
+  left: 0;
+  font-weight: 400;
+  font-size: 12px;
+  color: var(--danger);
+  &::first-letter {
+    text-transform: uppercase;
+  }
+}
+
+.radio-wrapper {
+  position: relative;
+  font-weight: 400;
+  font-size: 12px;
+  color: var(--black-4);
+
+  &.has-error {
+    color: var(--danger);
+
+    .input-error {
+      display: block;
+    }
+  }
+
+  & > div {
+    width: 100%;
+    display: flex;
+    gap: 40px;
+    margin: 10px 0;
+  }
+
+  input[type="radio"] {
+    display: none;
+  }
+
+  label {
+    display: flex;
+    align-items: center;
+    justify-content: flex-start;
+    gap: 16px;
+    font-weight: 400;
+    font-size: 14px;
+    color: var(--black-5);
+    cursor: pointer;
+
+    &::before {
+      content: "";
+      flex-shrink: 0;
+      display: block;
+      width: 24px;
+      height: 24px;
+      border-radius: 24px;
+      border: 1px solid var(--black-1);
+    }
+
+    &:has(input:checked) {
+      &::before {
+        border: 1px solid var(--black-4);
+        background: radial-gradient(var(--black-4) 55%, #ffffff 60%, #ffffff);
+      }
+    }
+  }
+}
+
+.list {
+  position: absolute;
+  top: 100%;
+  left: 0;
+  right: 0;
+  height: 200px;
+  border: 1px solid var(--black-4);
+  border-radius: 0 0 8px 8px;
+  scrollbar-width: none;
+  z-index: 5;
+  background: #ffffff;
+  &::-webkit-scrollbar {
+    width: 0;
+  }
+
+  &__option {
+    padding: 16px;
+    font-weight: 400;
+    font-size: 16px;
+    color: var(--black-5);
+    cursor: pointer;
+
+    &:hover {
+      background: var(--black-1);
+    }
+  }
+}
+
+.checkbox-wrapper {
+  position: relative;
+  input[type="checkbox"] {
+    display: none;
+  }
+
+  &.has-error {
+    color: var(--danger);
+
+    .input-error {
+      display: block;
+    }
+  }
+
+  label {
+    display: flex;
+    align-items: center;
+    justify-content: flex-start;
+    gap: 16px;
+    font-weight: 400;
+    font-size: 14px;
+    text-align: right;
+    color: var(--black-5);
+    cursor: pointer;
+
+    &::before {
+      content: "";
+      flex-shrink: 0;
+      display: block;
+      width: 24px;
+      height: 24px;
+      border-radius: 8px;
+      border: 1px solid var(--black-1);
+    }
+
+    &:has(input:checked) {
+      &::before {
+        border: 1px solid var(--black-4);
+        background: url("data:image/svg+xml,%3Csvg width='20' height='20' viewBox='0 0 20 20' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Crect width='20' height='20' rx='6' fill='%235e6366'/%3E%3Cpath d='M14 7L8.5 12.5L6 10' stroke='%23ffffff' stroke-width='2' stroke-linecap='round' stroke-linejoin='round'/%3E%3C/svg%3E%0A")
+          no-repeat center;
+      }
+    }
+  }
+}
+
+.send {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  max-width: 400px;
+  margin: 0 auto;
+  border: none;
+  border-radius: 12px;
+  padding: 17px 16px;
+  font-weight: 400;
+  font-size: 20px;
+  text-align: center;
+  color: #fff;
+  background: var(--black-3);
+  cursor: pointer;
+  transition: background-color 0.5s;
+
+  &:disabled {
+    pointer-events: none;
+    opacity: 0.5;
+  }
+
+  &.success {
+    background: var(--green);
+  }
+}
+
+.preview {
+  img {
+    width: 150px;
+    height: 150px;
+    margin-top: 8px;
+    display: block;
+    object-fit: cover;
+    object-position: center;
+    border-radius: 16px;
+    overflow: hidden;
+  }
+}
+
+.password-straight-wrapper {
+  position: absolute;
+  inset: 0;
+  border-radius: 7px;
+  overflow: hidden;
+  z-index: -1;
+}
+.password-straight {
+  position: absolute;
+  inset: 0 auto auto 0;
+  width: 0%;
+  height: 6px;
+  background: linear-gradient(
+    90deg,
+    #f57e77 0px,
+    #f57e77 120px,
+    #f5a477 120px,
+    #f5a477 240px,
+    #f5ca77 240px,
+    #f5ca77 360px,
+    #f2f577 360px,
+    #f2f577 480px,
+    #74df71 480px,
+    #74df71 600px
+  );
+  transition: width 0.5s;
+}
