+++ src/pages/uncontrolled/ui/Uncontrolled.tsx
@@ -1,88 +1,28 @@
 import { useRef, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { nanoid } from "nanoid";
-import {
-  InferType,
-  mixed,
-  number,
-  object,
-  ref,
-  string,
-  ValidationError,
-} from "yup";
+import { ValidationError } from "yup";
 
-import { countries } from "@/pages/controlled/model/constants";
 import { useAppDispatch, useAppSelector } from "@/shared/store";
 import { Eye, EyeOff } from "@/shared/ui/icons";
 
-import { addLastUpdated, addToSubmitHistory } from "../model";
-
-import "./Uncontrolled.scss";
-
-const MAX_FILE_SIZE = 2097152;
-
-const validFileExtensions: { [key: string]: string[] } = {
-  image: ["jpg", "png", "jpeg"],
-};
-
-const isValidFileType = (fileName: string, fileType: string) => {
-  const extension = fileName.split(".").pop();
-  return (
-    extension !== undefined &&
-    validFileExtensions[fileType]?.indexOf(extension) > -1
-  );
-};
-
-const userSchema = object({
-  name: string()
-    .required()
-    .matches(/^[A-Z]/, "Name must start with capital letter"),
-  age: number().required().positive().integer().typeError("Age is required"),
-  email: string()
-    .required()
-    .email()
-    .matches(
-      /^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$/,
-      "Enter valid email",
-    ),
-  gender: string()
-    .required()
-    .oneOf(["male", "female", "other"], "Choose your gender"),
-  country: string().required().oneOf(countries, "the country must be real"),
-  terms: string()
-    .required()
-    .oneOf(["on"], "You must accept the terms and conditions"),
-  file: mixed()
-    .required("Avatar required")
-    .test("is-valid-type", "Not a valid image type", (value) => {
-      if (!value || !(value instanceof File)) {
-        return false;
-      }
-      return isValidFileType(value.name.toLowerCase(), "image");
-    })
-    .test("is-valid-size", "Max allowed size is 2MB", (value) => {
-      if (!value || !(value instanceof File)) {
-        return false;
-      }
-      return value.size <= MAX_FILE_SIZE;
-    }),
-  password: string()
-    .required()
-    .matches(/[A-Z]/g, "Password must have 1 uppercase letter")
-    .matches(/[a-z]/g, "Password must have 1 lowercase letter")
-    .matches(/[0-9]/g, "Password must have 1 number")
-    .matches(/[!@#$%^&*]/g, "Password must have 1 special character")
-    .min(8),
-  passwordConfirm: string()
-    .required("Be sure to confirm your password")
-    .oneOf([ref("password")], "Passwords must match"),
-});
-
-export type FormData = InferType<typeof userSchema>;
+import { addLastUpdated, addToSubmitHistory, userSchema } from "../model";
 
 interface ErrorsObject {
   [key: string]: string;
 }
+
+const emptyErrors: ErrorsObject = {
+  name: "",
+  age: "",
+  email: "",
+  gender: "",
+  password: "",
+  passwordConfirm: "",
+  country: "",
+  terms: "",
+  file: "",
+};
 
 const Uncontrolled = () => {
   const [preview, setPreview] = useState<string | null>(null);
@@ -90,17 +30,8 @@ const Uncontrolled = () => {
   const countries = useAppSelector((state) => state.countries);
   const [isVisible, setIsVisible] = useState(false);
   const [indicatorWidth, setIndicatorWidth] = useState(0);
-  const [errors, setErrors] = useState<ErrorsObject>({
-    name: "",
-    age: "",
-    email: "",
-    gender: "",
-    password: "",
-    passwordConfirm: "",
-    country: "",
-    terms: "",
-    file: "",
-  });
+  const [errors, setErrors] = useState<ErrorsObject>(emptyErrors);
+  const [isSuccess, setIsSuccess] = useState(false);
 
   const form = useRef<HTMLFormElement | null>(null);
   const fileInputRef = useRef<HTMLInputElement | null>(null);
@@ -176,7 +107,12 @@ const Uncontrolled = () => {
 
             dispatch(addLastUpdated(temp.id));
             dispatch(addToSubmitHistory(temp));
-            navigate("/");
+
+            setIsSuccess(true);
+
+            setTimeout(() => {
+              navigate("/");
+            }, 1000);
           };
 
           reader.readAsDataURL(file);
@@ -184,15 +120,7 @@ const Uncontrolled = () => {
       } catch (error) {
         if (error instanceof ValidationError) {
           const eTemp: ErrorsObject = {
-            name: "",
-            age: "",
-            email: "",
-            gender: "",
-            password: "",
-            passwordConfirm: "",
-            country: "",
-            terms: "",
-            file: "",
+            ...emptyErrors,
           };
 
           error.inner.forEach((item: ValidationError) => {
@@ -357,7 +285,12 @@ const Uncontrolled = () => {
             <span className="input-error">{errors.terms}</span>
           </div>
 
-          <button className="send">Send</button>
+          <button
+            className={`send ${isSuccess ? "success" : ""}`}
+            type="submit"
+          >
+            Send
+          </button>
         </form>
       </div>
     </main>
